requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asgiref==3.11.0
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
bidict==0.23.1
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
cloudinary==1.44.1
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==6.0
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
h11==0.16.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
multidict==6.7.0
packaging==25.0
propcache==0.4.1
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-engineio==4.12.3
python-socketio==5.15.0
redis==7.1.0
service-identity==24.2.0
simple-websocket==1.1.0
six==1.17.0
sqlparse==0.5.4
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.1
wsproto==1.3.2
yarl==1.22.0
zope.interface==8.1.1
